# Conflicts:
#	app/src/main/java/com/kitzapp/telegram_stats/Application/AppManagers/ThemeManager.java
#	app/src/main/java/com/kitzapp/telegram_stats/presentation/ui/components/simple/TChartTextView.java
#	app/src/main/java/com/kitzapp/telegram_stats/presentation/ui/components/simple/TDelimiterLine.java

+++ app/src/main/java/com/kitzapp/telegram_stats/presentation/ui/components/TFullCellView.java
@@ -83,7 +83,7 @@ public class TFullCellView extends LinearLayout implements TViewObserver {
 
     @Override
     public void update(Observable o, Object arg) {
-        if ((int) arg == ObserverManager.KEY_OBSERVER_THEME_UPDATED) {
+        if ((byte) arg == ObserverManager.KEY_OBSERVER_THEME_UPDATED) {
             int newBackColor = getCurrentBackColor();
 
             if (_oldBackColor != newBackColor) {
