+++ app/src/main/java/com/kitzapp/telegram_stats/presentation/ui/components/simple/TColorfulTextView.java
@@ -62,7 +62,7 @@ public class TColorfulTextView extends TTextView implements TViewObserver {
 
     @Override
     public void update(Observable o, Object arg) {
-        if ((int) arg == ObserverManager.KEY_OBSERVER_THEME_UPDATED) {
+        if ((byte) arg == ObserverManager.KEY_OBSERVER_THEME_UPDATED) {
             int newColor = getCurrentColor();
             if (_oldTextColor != newColor) {
                 ValueAnimator valueAnimator = AndroidUtilites.getArgbAnimator(
